NXP-21651: add new implementation of tag service based on facet
