+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-api/src/main/java/org/nuxeo/ecm/platform/tag/TagService.java
@@ -20,6 +20,7 @@
 package org.nuxeo.ecm.platform.tag;
 
 import java.util.List;
+import java.util.Set;
 
 import org.nuxeo.ecm.core.api.CoreSession;
 
@@ -48,8 +49,20 @@ public interface TagService {
      * @param session the session
      * @param docId the document id
      * @param label the tag
-     * @param username the user associated to the tagging
+     * @since 9.3
      */
+    void tag(CoreSession session, String docId, String label);
+
+    /**
+     * Tags a document with a given tag.
+     *
+     * @param session the session
+     * @param docId the document id
+     * @param label the tag
+     * @param username the user associated to the tagging
+     * @deprecated since 9.3, username is not needed anymore
+     */
+    @Deprecated
     void tag(CoreSession session, String docId, String label, String username);
 
     /**
@@ -58,8 +71,19 @@ public interface TagService {
      * @param session the session
      * @param docId the document id
      * @param label the tag, or {@code null} for all tags
-     * @param username the user associated to the tagging
      */
+    void untag(CoreSession session, String docId, String label);
+
+    /**
+     * Untags a document of the given tag
+     *
+     * @param session the session
+     * @param docId the document id
+     * @param label the tag, or {@code null} for all tags
+     * @param username the user associated to the tagging
+     * @deprecated since 9.3, username is not needed anymore
+     */
+    @Deprecated
     void untag(CoreSession session, String docId, String label, String username);
 
     /**
@@ -74,13 +98,25 @@ public interface TagService {
     boolean canUntag(CoreSession session, String docId, String label);
 
     /**
+     * Gets the tags applied to a document.
+     *
+     * @param session the session
+     * @param docId the document id
+     * @return the list of tags
+     * @since 9.3
+     */
+    Set<String> getTags(CoreSession session, String docId);
+
+    /**
      * Gets the tags applied to a document by a given user, or by all users.
      *
      * @param session the session
      * @param docId the document id
      * @param username the user name, or {@code null} for all users
      * @return the list of tags
+     * @deprecated since 9.3, username is not needed anymore
      */
+    @Deprecated
     List<Tag> getDocumentTags(CoreSession session, String docId, String username);
 
     /**
@@ -94,7 +130,9 @@ public interface TagService {
      * @param useCore if true, the core should be used to retrieve tags.
      * @return the list of tags
      * @since 6.0
+     * @deprecated since 9.3, username and useCore are not needed anymore
      */
+    @Deprecated
     List<Tag> getDocumentTags(CoreSession session, String docId, String username, boolean useCore);
 
     /**
@@ -129,9 +167,21 @@ public interface TagService {
      *
      * @param session the session
      * @param label the tag
+     * @return the set of document ids
+     * @since 9.3
+     */
+    List<String> getTagDocumentIds(CoreSession session, String label);
+
+    /**
+     * Gets the documents to which a tag is applied.
+     *
+     * @param session the session
+     * @param label the tag
      * @param username the user name, or {@code null} for all users
      * @return the set of document ids
+     * @deprecated since 9.3, username is not needed anymore
      */
+    @Deprecated
     List<String> getTagDocumentIds(CoreSession session, String label, String username);
 
     /**
@@ -147,8 +197,20 @@ public interface TagService {
      * @param normalize null for no weight normalization (a count is returned), {@code FALSE} for 0-100 normalization,
      *            {@code TRUE} for logarithmic 0-100 normalization
      * @return the cloud (a list of weighted tags)
+     * @deprecated since 9.3, seems unused
      */
+    @Deprecated
     List<Tag> getTagCloud(CoreSession session, String docId, String username, Boolean normalize);
+
+    /**
+     * Gets suggestions for a given tag label prefix.
+     *
+     * @param session the session
+     * @param label the tag label prefix
+     * @return a list of tags
+     * @since 9.3
+     */
+    Set<String> getSuggestions(CoreSession session, String label);
 
     /**
      * Gets suggestions for a given tag label prefix.
@@ -157,7 +219,9 @@ public interface TagService {
      * @param label the tag label prefix
      * @param username the user name, or {@code null} for all users
      * @return a list of tags
+     * @deprecated since 9.3, username is not needed anymore
      */
+    @Deprecated
     List<Tag> getSuggestions(CoreSession session, String label, String username);
 
 }
