+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-core/src/main/java/org/nuxeo/ecm/platform/tag/TaggedVersionListener.java
@@ -65,10 +65,6 @@ public class TaggedVersionListener implements PostCommitFilteringEventListener {
             }
             String docId = doc.getId();
             TagService tagService = Framework.getLocalService(TagService.class);
-            if (doc instanceof DeletedDocumentModel) {
-                tagService.removeTags(session, docId);
-                return;
-            }
             switch (name) {
             case DOCUMENT_PROXY_PUBLISHED:
                 if (doc.isProxy()) {
@@ -80,7 +76,6 @@ public class TaggedVersionListener implements PostCommitFilteringEventListener {
                 String versionUUID = (String) ctx.getProperty(VersioningDocument.RESTORED_VERSION_UUID_KEY);
                 tagService.replaceTags(session, versionUUID, docId);
                 break;
-            case DOCUMENT_REMOVED:
             case TRANSITION_EVENT:
                 tagService.removeTags(session, docId);
                 break;
@@ -93,10 +88,8 @@ public class TaggedVersionListener implements PostCommitFilteringEventListener {
     @Override
     public boolean acceptEvent(Event event) {
         String name = event.getName();
-        return DOCUMENT_PROXY_PUBLISHED.equals(name)
-                || DOCUMENT_RESTORED.equals(name)
-                || DOCUMENT_REMOVED.equals(name)
-                || (LifeCycleConstants.TRANSITION_EVENT.equals(name) && LifeCycleConstants.DELETED_STATE.equals(event.getContext().getProperty(
-                        LifeCycleConstants.TRANSTION_EVENT_OPTION_TO)));
+        return DOCUMENT_PROXY_PUBLISHED.equals(name) || DOCUMENT_RESTORED.equals(name)
+                || (LifeCycleConstants.TRANSITION_EVENT.equals(name) && LifeCycleConstants.DELETED_STATE.equals(
+                        event.getContext().getProperty(LifeCycleConstants.TRANSTION_EVENT_OPTION_TO)));
     }
 }
