+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-core/src/main/resources/META-INF/MANIFEST.MF
@@ -18,7 +18,9 @@ Nuxeo-Component: OSGI-INF/TagService.xml,
   OSGI-INF/ws-tag-loader-contrib.xml,
   OSGI-INF/tag-pageprovider-contrib.xml,
   OSGI-INF/tag-operations-contrib.xml,
-  OSGI-INF/json-enrichers-contrib.xml
+  OSGI-INF/json-enrichers-contrib.xml,
+  OSGI-INF/faceted-tag-service-core-types.xml,
+  OSGI-INF/tag-service-properties.xml
 Bundle-ManifestVersion: 2
 Import-Package: javax.resource,
  org.apache.commons.logging,
