+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-core/src/main/resources/OSGI-INF/faceted-tag-service-core-types.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0"?>
+
+<component name="org.nuxeo.ecm.tags.schemas">
+
+  <require>org.nuxeo.ecm.core.CoreExtensions</require>
+
+  <extension target="org.nuxeo.ecm.core.schema.TypeService" point="schema">
+
+    <schema name="facetedTag" src="schemas/facetedTag.xsd" prefix="nxtag" />
+
+  </extension>
+
+  <extension target="org.nuxeo.ecm.core.schema.TypeService" point="doctype">
+
+    <facet name="Tag">
+      <schema name="facetedTag" />
+    </facet>
+
+    <doctype name="File" append="true">
+      <facet name="Tag" />
+    </doctype>
+
+    <doctype name="Note" append="true">
+      <facet name="Tag" />
+    </doctype>
+
+    <doctype name="Picture" append="true">
+      <facet name="Tag" />
+    </doctype>
+
+    <doctype name="Video" append="true">
+      <facet name="Tag" />
+    </doctype>
+
+    <doctype name="Sound" append="true">
+      <facet name="Tag" />
+    </doctype>
+
+  </extension>
+
+</component>
