+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-core/src/main/resources/OSGI-INF/tag-service-properties.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<component name="org.nuxeo.tag.service.properties">
+
+  <extension target="org.nuxeo.runtime.ConfigurationService" point="configuration">
+    <documentation>
+      Property that enables choosing between the former (based on SQL relations) or the new (based on facet) tag service
+      implementation.
+    </documentation>
+    <property name="nuxeo.faceted.tag.service.enabled">true</property>
+  </extension>
+
+</component>
