+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-core/src/main/resources/schemas/facetedTag.xsd
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<xs:schema targetNamespace="http://www.nuxeo.org/ecm/schemas/facetedTag/" xmlns:xs="http://www.w3.org/2001/XMLSchema"
+  xmlns:tagxs="http://www.nuxeo.org/ecm/schemas/facetedTag/">
+
+  <xs:complexType name="stringList">
+    <xs:sequence>
+      <xs:element name="tag" type="xs:string" minOccurs="0" maxOccurs="unbounded"/>
+    </xs:sequence>
+  </xs:complexType>
+
+  <xs:element name="tags" type="tagxs:stringList"/>
+
+</xs:schema>
