+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-core/src/test/java/org/nuxeo/ecm/platform/tag/TestRelationTagService.java
@@ -1,0 +1,32 @@
+/*
+ * (C) Copyright 2017 Nuxeo (http://nuxeo.com/) and others.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Contributors:
+ *     Funsho David
+ *
+ */
+
+package org.nuxeo.ecm.platform.tag;
+
+import org.nuxeo.runtime.test.runner.LocalDeploy;
+
+/**
+ * Test class for tag service based on SQL relations
+ *
+ * @since 9.3
+ */
+@LocalDeploy("org.nuxeo.ecm.platform.tag:override-tag-service-properties.xml")
+public class TestRelationTagService extends AbstractTestTagService {
+}
