+++ nuxeo-features/nuxeo-platform-tag-service/nuxeo-platform-tag-core/src/test/resources/override-tag-service-properties.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<component name="org.nuxeo.tag.service.properties.override">
+
+  <require>org.nuxeo.tag.service.properties</require>
+
+  <extension target="org.nuxeo.runtime.ConfigurationService" point="configuration">
+    <property name="nuxeo.faceted.tag.service.enabled">false</property>
+  </extension>
+
+</component>
